migrate(tests): merge MLX segmentation tests into test_mlx_inference.py
Consolidates segmentation inference tests into `test_mlx_inference.py` for unified MLX test coverage. Removes the now-redundant `test_mlx_seg_inference.py`. Updates file-level docstrings to include segmentation tests.
